modularize search algo
