+++ app/api/untappd.js
@@ -1,10 +1,13 @@
 /*
     Todo:
-    > data.response.beers.items[0].brewery.location
-    {brewery_city: 'Chicago',
-    brewery_state: 'IL',
-    lat: 41.8871,
-    lng: -87.6721 }
+
+    Beer search to get BID
+    data.response.beers.items[0].beer.bid for first
+
+    Modularize search similar to brewerydb
+
+    Use Beer info to get venue with the BID
+    Search for closest long-lat according to user location
 */
 
 
@@ -12,9 +15,10 @@
 var utdApi = {
 	'parseResp':
 		function(beer, response) {
-
-			// Todo
-		  return response;
+            var data = JSON.parse(response);
+            // debugger;
+            // Todo
+            return data;
 		}
 }
